refactor: reorganize Vue 3 project structure
