+++ src/components/sections/SectionFooter.vue
@@ -1,5 +1,5 @@
 <script setup>
-import VText from './ui/VText.vue';
+import VText from '../ui/VText.vue';
 
 
 </script>
