+++ src/components/sections/SectionProjects.vue
@@ -1,9 +1,11 @@
 <script setup>
-import VCard from '../components/ui/VCard.vue';
-import ImgMobile from "@/assets/mobile.jpg"
-import ImgWeb from '@/assets/web.jpg'
-import ImgDashboard from '@/assets/dashboard.jpg'
-import VText from './ui/VText.vue';
+import VCard from '../ui/VCard.vue';
+import VText from '../ui/VText.vue';
+
+import ImgMobile from "@/assets/images/mobile.jpg"
+import ImgWeb from '@/assets/images/web.jpg'
+import ImgDashboard from '@/assets/images/dashboard.jpg'
+
 const items = [
   { id: 1, name: 'Website responsivo', img: ImgWeb, description: 'UI limpa • Acessível • Responsiva' },
   { id: 2, name: 'Aplicativo mobile UI', img: ImgMobile, description: 'UI limpa • Acessível • Responsiva' },
