+++ src/main.ts
@@ -1,4 +1,4 @@
-import './main.css'
+import './assets/styles/main.css'
 
 import { createApp } from 'vue'
 import App from './App.vue'
